Sim2Real adhoc bugfixes (#113)
* Added QoL to grasp_eval: tqdm

* Bugfix: new Genesis changed the role of VisOptions.rendered_envs_idx parameter

* Adhoc bugfixes and todos

* Getting config changes from ClearML (#94)

* Getting config changes from ClearML


* Train config makeover

* added logging of the dataset

* Applied new GraspConfig object to train and eval scripts

* formatting

* Added tqdm to grasp

* updated readme

* Added flag for disabling cfg logging in rsl_rl

* Changed all lauch arguments to kebab-case

* Cleanup after merging of the sweep evaluation script

* Added ability to reset RL/BC last layers (#93)

* Implemented reseting last layers weights for RL and BC

* Updated changelog

* Added todos; config reset

* Updated pre-commit

* Changed default RL networks from 256 to 128

* updated pre-commit

* added TODO comments

* dumping all config

* typos

* Improved typing

* typo

* added test script for running train and eval

* Resolving RL/BC config from clearml (pickles are not longer needed)

* typo

* Added control over loading checkpoitn cfg from clearml

* more logging

* catching the KeyboardInterrupt

* improved logging

* Updated logs

---------


* WIP on sim2real

* Getting config changes from ClearML (#94)

* Getting config changes from ClearML


* Train config makeover

* added logging of the dataset

* Applied new GraspConfig object to train and eval scripts

* formatting

* Added tqdm to grasp

* updated readme

* Added flag for disabling cfg logging in rsl_rl

* Changed all lauch arguments to kebab-case

* Cleanup after merging of the sweep evaluation script

* Added ability to reset RL/BC last layers (#93)

* Implemented reseting last layers weights for RL and BC

* Updated changelog

* Added todos; config reset

* Updated pre-commit

* Changed default RL networks from 256 to 128

* updated pre-commit

* added TODO comments

* dumping all config

* typos

* Improved typing

* typo

* added test script for running train and eval

* Resolving RL/BC config from clearml (pickles are not longer needed)

* typo

* Added control over loading checkpoitn cfg from clearml

* more logging

* catching the KeyboardInterrupt

* improved logging

* Updated logs

---------


* typos in calling eval_policy methods

* Working sim2real transfer

* pre-commit formatting

* Restored config

* Added argument for overwriting the episode-length

* Prepared TODO

* BC cleanup

* Added debug option for swapping tool cameras

* cleanup

* fixed typos

* Apply suggestions from code review


* Update aegis_gym/rsl/envs/grasp_env_ros.py

* pre-commit formatting

---------

+++ aegis_gym/rsl/envs/manipulator.py
@@ -303,7 +303,7 @@ class Manipulator:
 
     @property
     def base_pos(self):
-        return self._robot_entity.get_pos()
+        return self._robot_entity.get_pos().float()
 
     @property
     def ee_pose(self) -> th.Tensor:
@@ -311,7 +311,7 @@ class Manipulator:
         The end-effector pose (the hand pose)
         """
         pos, quat = self._ee_link.get_pos(), self._ee_link.get_quat()
-        return th.cat([pos, quat], dim=-1)
+        return th.cat([pos, quat], dim=-1).float()
 
     @property
     def gripper_width(self) -> th.Tensor:
